Secrets deployment in ssl modes 'self-signed' and 'issued'.

+++ scripts/check-issued-certs.sh
@@ -13,12 +13,14 @@
 
 echo "$(date +"%Y-%m-%d %H:%M:%S.%6N") - ${0}"
 
-cd "${AG_SOURCE_DIR}/platform/secrects/certs/issued" || exit 1
+cd "${AG_SOURCE_DIR}/platform/secrets/certs/issued" || exit 1
 
 # Rename any *.crt and *.key files to start with AG_PROJECT_BASE_URL
-mv ./*.crt "${AG_PROJECT_BASE_URL}.crt"
-mv ./*.key "${AG_PROJECT_BASE_URL}.key"
+[[ ! -f "${AG_PROJECT_BASE_URL}.crt" ]] && mv ./*.crt "${AG_PROJECT_BASE_URL}.crt"
+[[ ! -f "${AG_PROJECT_BASE_URL}.key" ]] && mv ./*.key "${AG_PROJECT_BASE_URL}.key"
 
 # Check if issued certificate and key with AG_PROJECT_BASE_URL as prefix exist.
-test -z "$(find . -maxdepth 1 -name "${AG_PROJECT_BASE_URL}.crt" -print -quit)" && exit 2
-test -z "$(find . -maxdepth 1 -name "${AG_PROJECT_BASE_URL}.key" -print -quit)" && exit 3
+if [[ ! -f "${AG_PROJECT_BASE_URL}.crt" ]] || [[ ! -f "${AG_PROJECT_BASE_URL}.key" ]]; then
+    >&2 echo "Issued certificate or key not found in directory '$(pwd)'."
+    exit 2
+fi
